LayoutTests/imported/w3c/web-platform-tests/loading/early-hints/preload-fetchpriority.h2.window-expected.txt

CONSOLE MESSAGE: Error: assert_true: expected true got false
CONSOLE MESSAGE: The resource https://localhost:9000/loading/early-hints/resources/empty.js?d6647a78-4d15-4278-800d-a6e0c99c8e37high was preloaded using link preload but not used within a few seconds from the window's load event. Please make sure it wasn't preloaded for nothing.
CONSOLE MESSAGE: The resource https://localhost:9000/loading/early-hints/resources/empty.js?66773332-a80b-4e87-8162-7b30fc6faee0auto was preloaded using link preload but not used within a few seconds from the window's load event. Please make sure it wasn't preloaded for nothing.
FAIL: Timed out waiting for notifyDone to be called
